+++ Runtime/Installers/IF_Installer.cs
@@ -15,6 +15,24 @@ namespace InterVR.IF.Installer
         public override void InstallBindings()
         {
             Container.BindInstance(settings).IfNotBound();
+
+            checkInstaller("intervr.if.vr");
+            checkInstaller("intervr.if.glove");
+        }
+
+        void checkInstaller(string path)
+        {
+            var prefab = Resources.Load(path) as GameObject;
+            if (prefab == null)
+                return;
+
+            Debug.Log($"found {path} package to install");
+
+            var instance = GameObject.Instantiate(prefab.gameObject);
+            instance.transform.SetParent(ProjectContext.Instance.transform, false);
+            var installer = instance.GetComponent<MonoInstaller>();
+            Container.Inject(installer);
+            installer.InstallBindings();
         }
 
         [Serializable]